refactor: decouple hubspot from download error handling
